[Fix] Fix timer logic
